remove hard data in my code, using mock data
