+++ competency_matrix/templates/download.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
   </head>
   <body>
-    <h1>Télécharger le zip file</h1>
-    <a href="/download/" download>Télécharger le zip file</a>
+    <h1>Télécharger le fichier zip</h1>
+    <a href="/download/" download>Télécharger</a>
   </body>
 </html>
